Move unit test from parse to error
In preparation for deleting the `parse` module (and clobbering the
`FromHex` trait) move the unit test to `error` so its obvious we don't
loose any test coverage.
