+++ src/error.rs
@@ -410,7 +410,8 @@ if_std_error! {{
 #[cfg(feature = "std")]
 mod tests {
     use super::*;
-    use crate::decode_to_array;
+    use crate::error::{InvalidCharError, OddLengthStringError};
+    use crate::{decode_to_array, HexToBytesIter, InvalidLengthError};
     #[cfg(feature = "alloc")]
     use crate::decode_to_vec;
 
@@ -478,5 +479,89 @@ mod tests {
         });
         assert!(!format!("{}", error).is_empty());
         check_source(&error);
+    }
+
+    #[test]
+    #[cfg(feature = "alloc")]
+    fn hex_error() {
+        let oddlen = "0123456789abcdef0";
+        let badchar1 = "Z123456789abcdef";
+        let badchar2 = "012Y456789abcdeb";
+        let badchar3 = "«23456789abcdef";
+
+        assert_eq!(
+            decode_to_vec(oddlen).unwrap_err(),
+            OddLengthStringError { len: 17 }.into()
+        );
+        assert_eq!(
+            decode_to_array::<4>(oddlen).unwrap_err(),
+            InvalidLengthError { invalid: 17, expected: 8 }.into()
+        );
+        assert_eq!(
+            decode_to_vec(badchar1).unwrap_err(),
+            InvalidCharError { pos: 0, invalid: b'Z' }.into()
+        );
+        assert_eq!(
+            decode_to_vec(badchar2).unwrap_err(),
+            InvalidCharError { pos: 3, invalid: b'Y' }.into()
+        );
+        assert_eq!(
+            decode_to_vec(badchar3).unwrap_err(),
+            InvalidCharError { pos: 0, invalid: 194 }.into()
+        );
+    }
+
+    #[test]
+    fn hex_error_position() {
+        let badpos1 = "Z123456789abcdef";
+        let badpos2 = "012Y456789abcdeb";
+        let badpos3 = "0123456789abcdeZ";
+        let badpos4 = "0123456789abYdef";
+
+        assert_eq!(
+            HexToBytesIter::new(badpos1).unwrap().next().unwrap().unwrap_err(),
+            InvalidCharError { pos: 0, invalid: b'Z' }
+        );
+        assert_eq!(
+            HexToBytesIter::new(badpos2).unwrap().nth(1).unwrap().unwrap_err(),
+            InvalidCharError { pos: 3, invalid: b'Y' }
+        );
+        assert_eq!(
+            HexToBytesIter::new(badpos3).unwrap().next_back().unwrap().unwrap_err(),
+            InvalidCharError { pos: 15, invalid: b'Z' }
+        );
+        assert_eq!(
+            HexToBytesIter::new(badpos4).unwrap().nth_back(1).unwrap().unwrap_err(),
+            InvalidCharError { pos: 12, invalid: b'Y' }
+        );
+    }
+
+    #[test]
+    fn hex_to_array() {
+        let len_sixteen = "0123456789abcdef";
+        assert!(decode_to_array::<8>(len_sixteen).is_ok());
+    }
+
+    #[test]
+    fn hex_to_array_error() {
+        let len_sixteen = "0123456789abcdef";
+        assert_eq!(
+            decode_to_array::<4>(len_sixteen).unwrap_err(),
+            InvalidLengthError { invalid: 16, expected: 8 }.into()
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "alloc")]
+    fn mixed_case() {
+        use crate::display::DisplayHex as _;
+
+        let s = "DEADbeef0123";
+        let want_lower = "deadbeef0123";
+        let want_upper = "DEADBEEF0123";
+
+        let v = decode_to_vec(s).expect("valid hex");
+        assert_eq!(format!("{:x}", v.as_hex()), want_lower);
+        assert_eq!(format!("{:X}", v.as_hex()), want_upper);
     }
 }
